+++ src/main/java/com/uade/tpo/demo/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.uade.tpo.demo.service;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.springframework.data.domain.Page;
@@ -16,11 +17,11 @@ public interface ProductService {
     
     public Page<Product> getProducts(PageRequest pageRequest);
     public Optional<Product> getProductById(Long productId);
-    public Product createProduct(String name, String description, Category category, Double price, Integer stock, String imageUrl, Double descuento) throws ProductDuplicateException;
+    public Product createProduct(String name, String description, Category category, BigDecimal price, Integer stock, String imageUrl, BigDecimal descuento) throws ProductDuplicateException;
     public Page<Product> getProductsByCategory(Long categoryId, PageRequest pageRequest);
-    public Product updateProduct(Long productId, String name, String description, Category category, Double price, Integer stock, String imageUrl, Double descuento) throws ProductNotFoundException;
+    public Product updateProduct(Long productId, String name, String description, Category category, BigDecimal price, Integer stock, String imageUrl, BigDecimal descuento) throws ProductNotFoundException;
     public void deleteProduct(Long productId) throws ProductNotFoundException;
-    public Page<Product> getProductsByPriceRange(Double minPrice, Double maxPrice, PageRequest pageRequest) throws ProductNotFoundException;
+    public Page<Product> getProductsByPriceRange(BigDecimal minPrice, BigDecimal maxPrice, PageRequest pageRequest) throws ProductNotFoundException;
